==> 02 Scripts/1. Data import and descriptive analysis.ipynb <==
## Data import and descriptive analysis

#import libraries

import pandas as pd
import numpy as np
import os

#import the orders.csv file and print first 5 rows
path = r'/Users/patel/Documents/CF-Data Anaylst Course/Achievement-4/Instacart Basket Analysis-Oct 2023/'

var_list = ['order_id', 'user_id', 'order_number','order_dow', 'order_hour_of_day', 'days_since_prior_order' ]

df_orders = pd.read_csv(os.path.join(path, '02 Data/Orginal Data/4.3_orders_products/orders.csv'), usecols=var_list)

df_orders.head()

#print number of columns and rows in orders.csv
df_orders.shape

# import the products.csv file and print first 20 rows
df_prod = pd.read_csv(os.path.join(path, '02 Data/Orginal Data/4.3_orders_products/products.csv'), index_col=False)

df_prod.head(20)

# print last 35 rows of product file
df_prod.tail(35)

# print column names
df_prod.columns

#print the number of columns and rows
df_prod.shape

# Check datatypes
df_prod.dtypes

#### Notes:

Max aisle_id = 134

datatype of department_id = int64 (integer)
