==> 02 Scripts/2. Data Wrangling & Subsetting.ipynb <==
# Data Wrangling & Subestting

# import libraries

import pandas as pd
import numpy as np
import os

## Loading & Data Wrangling order.csv

path = r'/Users/patel/Documents/CF-Data Anaylst Course/Achievement-4/Instacart Basket Analysis-Oct 2023/02 Data/'

#load orders.csv

# load specific columns required and shrink numerical values for less memory consumption
var_list = ['order_id','user_id','order_number','order_dow','order_hour_of_day','days_since_prior_order']
dtypes = {'order_dow' : 'int8','order_hour_of_day': 'int8', 'days_since_prior_order':'float32', 'order_number': 'int8'}

df_ords = pd.read_csv(os.path.join(path, 'Orginal Data/4.3_orders_products/orders.csv'), usecols=var_list, dtype = dtypes )

df_ords.head()


# descriptive analytics of orders
df_ords.describe()

# info on dtypes for orders
df_ords.info()

### Check fo missing values in orders.csv

# check for missing values - 6.0%
df_ords.value_counts("days_since_prior_order", dropna=False, sort=True)

### Rename Columns

# rename columns

df_ords.rename(columns = {'order_dow': 'orders_days_of_week'}, inplace=True)
df_ords.head()

### Change column types

# change column types for order_id and user_id to be string

df_ords['order_id'] = df_ords['order_id'].astype('str')

df_ords.dtypes

## Loading & Data Wrangling products.csv

# load product.csv and reducde memory consumption by shrinking numerical values

dtypes_prods = {'aisle_id': 'int16', 'department_id': 'int8'}

df_prods = pd.read_csv(os.path.join(path, 'Orginal Data/4.3_orders_products/products.csv'), dtype=dtypes_prods, index_col=False)

df_prods.head()

# descriptive analytics of produtcs
df_prods.describe()

# data types of products columns
df_prods.info()

# checking prices, counts and looking for Nulls
df_prods.value_counts("prices", dropna=False)

# Converting product_id into string
df_prods['product_id'] = df_prods['product_id'].astype('str')

df_prods.dtypes

## Loading & Data Wrangling department.csv

# load department file and review first 5 rows
df_dep = pd.read_csv(os.path.join(path, 'Orginal Data/4.3_orders_products/departments.csv'), nrows=10)

df_dep.head()

### Transponse Department Data

# Transpose department dataframe

df_dep_t = df_dep.T
df_dep_t.reset_index() # adding index

# add header to the department df

# copying the first row of dept_t dataframe which has the header values
new_header = df_dep_t.iloc[0]

# copy data from row 1 onwards
df_dept_new = df_dep_t[1:]

#replace the column header with new header
df_dept_new.columns = new_header
df_dept_new.head(22)


### Create Data Dictionary of Departments

# create a data dictionary for departments (df_dept_new)

data_dic_dep = df_dept_new.to_dict('index')
data_dic_dep

### Subsetting products example

#filter product data that has department_id 19
df_snacks = df_prods[df_prods['department_id']==19]
df_snacks.head()

## Update Order File Columns

### Changed user_id to string value

## Step 1 : change user_id to string. Not reuqired to be numeric

df_ords['user_id'] = df_ords['user_id'].astype('str')
df_ords.dtypes

### Changed column name of 'order_number' to 'number_of_orders'

df_ords.rename(columns = {'order_number': 'number_of_orders'}, inplace=True)
df_ords.head()

## EDA 

### Find the busiest hour for placing orders

# Results below
# order_hour_of_day, order_count
# 10    288,418
# 11    284,728
# 15    283,639
# 14    283,042
# 13    277,999

df_ords['order_hour_of_day'].value_counts()

### Insights

1. Pattern: orders start to ramp up from 8am (178,201), peak at 10am (288,418), slows down for lunch hours 12:00-13:00, ramps up again at between 14-15 and start to slow down until 20:00

2. Peak ordering periods 10:00, 11:00 and 15:00

### Create a subset containing breakfast items sales

#print dictionary to find the key for breakfast
data_dic_dep

# 14 is the department_id for breakfast
# create a subset by filtering products with breakfast id = 14

df_breakfast_items = df_prods[df_prods['department_id']==14]
df_breakfast_items.head(10)

### Create subset of dinner party products

# Find products for dinner parties (food from departments alcohol, deli, beverages, and meat/seafood)

# '5': {'department': 'alcohol'},
# '7': {'department': 'beverages'},
# '12': {'department': 'meat seafood'},
# '20': {'department': 'deli'}

df_dinner_parties = df_prods.loc[df_prods['department_id'].isin([5,7,12,20])]
df_dinner_parties

### Number of rows in df_dinner_parties

# 7,650 rows (dinner party items)

df_dinner_parties.shape

### Create subset of orders made by user_id = 1

# Extract information about user id = 1

df_ords_userid_1 = df_ords[df_ords['user_id']=='1']
df_ords_userid_1

### Find basic stats and behaviours about user_id =1

# descriptive stats of user_id 1 orders

df_ords_userid_1.describe()

# when does the user_id normally order
df_ords_userid_1.value_counts('order_hour_of_day')

# most orders placed at 7 and 8am in the morning

# days of the week the user normally places orders
df_ords_userid_1.value_counts('orders_days_of_week')

# Most orders placed on Wednesday(4), followed by Sunday (1)

# max spend on groceries

df_ords_userid_1['days_since_prior_order'].max()

# 30.0

### Basic User ID 1 Behaviour

Below are some insights in behaviour

1. Normally places orders between 7am and 4pm. Most orders placed at 7am and 8am
2. Orders are normally placed on Sunday and Wednesday
3. Maximum spent on groccery is 30.0
4. Total of 11 orders made

### Export orders dataframe and place in “Prepared Data” folder

df_ords.to_csv(os.path.join(path, 'Prepared Data/orders_wrangled.csv'))

### Export departments dataframe and place in “Prepared Data” folder

df_dept_new

df_dept_new.to_csv(os.path.join(path,'Prepared Data/departments_wrangled.csv'))
