==> 02 Scripts/3. Data Consistency.ipynb <==
# Date Consistency Checks

# import libraries
import pandas as pd
import numpy as np
import os

## Products.csv

### Import file

# load product.csv

path = r'/Users/patel/Documents/CF-Data Anaylst Course/Achievement-4/Instacart Basket Analysis-Oct 2023/02 Data/'
dtypes = {'aisle_id': 'int16', 'department_id': 'int8'}
df_prods = pd.read_csv(os.path.join(path, 'Orginal Data/4.3_orders_products/products.csv'), dtype=dtypes, index_col=False)

#convert product_id to string

df_prods['product_id'] = df_prods['product_id'].astype('str')

df_prods.info()

### Missing Values

# find columns with missing values
df_prods.isnull().sum()

#### product_name -> 16 missing values ( 0.032%)

# find rows with missing product name

df_na_prod_name = df_prods[df_prods['product_name'].isnull() == True]
df_na_prod_name

#### Create new dataframe without missing product name rows

df_prods_clean = df_prods[df_prods['product_name'].isnull() == False]
df_prods_clean.shape

### Find and Manage Duplicates

#### Find duplicates in prods_clean

#find list of duplicated rows based on all column values - 5 rows
df_dups = df_prods_clean[df_prods_clean.duplicated()]

df_dups


#### Remove duplicates from prods

df_prods_clean_no_dups = df_prods_clean.drop_duplicates()
df_prods_clean_no_dups.shape

df_dups2 = df_prods_clean_no_dups[df_prods_clean_no_dups.duplicated(subset=['product_id'], keep=False)]
df_dups2

### Export Cleaned Products DF to CSV

df_prods_clean_no_dups.to_csv(os.path.join(path, 'Prepared Data/products_checked.csv'))

## Wrangled Orders File

### Import wrangled orders file

# To reduce memory, cast data types while importing file into dataframe
dtypes = {'order_number': 'int8', 'order_dow': 'int8', 'order_hour_of_day': 'int8','days_since_prior_order': 'float32' }

df_ords = pd.read_csv(os.path.join(path, 'Prepared Data/orders_wrangled.csv'), dtype=dtypes, index_col=0)

df_ords['order_id'] = df_ords['order_id'].astype('str')
df_ords['user_id'] = df_ords['user_id'].astype('str')
df_ords.info()

### Check for irregularities 

df_ords.describe()

#### Data Observations

days_since_prior order:
    1. Count of rows is less 206,209
    2. Max is 30 days. Is the data obtained for orders in last 30 days?

### Check of mixed data types

#code to check for columns that have mixed data type
for col in df_ords.columns.tolist():
    weird = (df_ords[[col]].map(type) != df_ords[[col]].iloc[0].apply(type)).any(axis = 1)
    
    if len (df_ords[weird]) > 0:
        print (col)

no mixed datype column found

### Check for Missing values

# find columns with missing values
df_ords.isnull().sum()

# check missing values in days_since_prior_order

df_days_since_missing = df_ords[df_ords['days_since_prior_order'].isnull()]
df_days_since_missing.value_counts('number_of_orders')

#### 206,209 values missing in days_since_prior_order column

# displaying rows with missing days_since_prior_order
df_days_since_missing

# Double check if any data with 1 'number of orders' and days_since_prior_order is not null
df_ords[(df_ords['number_of_orders']==1) & (df_ords['days_since_prior_order'].isnull()==False)]

crosstab = pd.crosstab(df_ords['days_since_prior_order'], df_ords['number_of_orders'], dropna = False)

crosstab.to_clipboard()

#### Address missing values

All users (206,209) have their first order a missing 'days_since_prior_order' value. This makes sense since this is their first order recorded and so having any value days_since_prior_order wouldn't make sense

I have decided to leave the data as is and not impute values. The data can help recognise customers who have only made 1 order or new customers. 

### Check & Address Duplicates

# create a subset of duplicate records and check shape
df_dups = df_ords[df_ords.duplicated()]
df_dups.shape

Number of rows df_dups = 0. This shows there are no duplicates in orders dataset

### Export orders

df_ords.to_csv(os.path.join(path, 'Prepared Data/orders_checked.csv'))
