==> 02 Scripts/4. Combine & Export - Part 2.ipynb <==
# Combine Datasets and Export

# import libraries

import pandas as pd
import numpy as np
import os

## Import Files

### Import Orders Products Combined Pickle File

# import the orders_products_combined.pkl file

path = r'/Users/patel/Documents/CF-Data Anaylst Course/Achievement-4/Instacart Basket Analysis-Oct 2023/02 Data/'
df_ords_merged = pd.read_pickle(os.path.join(path, 'Prepared Data/orders_products_combined.pkl'))
df_ords_merged.head()

### Check shape of dataframe

df_ords_merged.shape

df_ords_merged.info()

# drop the _merge column as its no longer required
df_ords_merged.drop(columns=['_merge'], inplace=True)

# check if the drop columb action was successful
df_ords_merged.info()

### Merge Products and Orders Product datasets

#### 1. Import products file

dtypes = {'aisle_id': 'int16', 'department_id': 'int8'}
df_prods = pd.read_csv(os.path.join(path, 'Prepared Data/products_checked.csv'), dtype=dtypes, index_col=0)

df_prods['product_id'] = df_prods['product_id'].astype('str') 
df_prods.head()

df_prods.shape

df_prods.info()

df_prods_dups = df_prods[df_prods.duplicated(subset=['product_id'], keep = False)]
df_prods_dups.head(20)

#### 2. Merge orders_products with products using left join

# used left join to combine both datasets
df_ords_prods_combined = df_ords_merged.merge(df_prods, on=['product_id'], how='left', indicator=True)
df_ords_prods_combined.head()

df_ords_prods_combined.shape

#### ---------

#### Investigating difference in row count between orders_product and merge database. Expected left merged dataset to have 32434489 rows. Instead extra 570 rows was in the combined set

# checking difference in rows
32435059 - 32434489

# checking for duplicated rows with combination of order_id and product_id
df_dups = df_ords_prods_combined[df_ords_prods_combined.duplicated(subset=['order_id', 'product_id'], keep='first')]
df_dups.shape

# found 2 product ids that are resulting in additional 570 rows
df_dups['product_id'].unique()

# same product id shows 2 different products
df_prods_dups = df_prods[df_prods.duplicated(subset=['product_id'], keep = False)]
df_prods_dups.head(20)

using left join, extra 570 rows appear in the combined set is because duplicate product_ids exists in product table. They are not unique.

### ----------

#### Confirmed list merge using merge flag

# checking the rows count of merge flag
df_ords_prods_combined.value_counts('_merge')

# checking on number of null values since we have left_only with 30,200 rows
df_ords_prods_combined.isnull().sum()

# extract list of product_ids in orders_products_combined but doesn't exists in product table
null_product_name = df_ords_prods_combined[df_ords_prods_combined['product_name'].isnull()]
null_product_name_prod_id = null_product_name['product_id'].unique()
null_product_name_prod_id.shape

df_prods[(df_prods['product_id'].isin(null_product_name_prod_id) == True)]

#### found 18 products that is not in products table. Ideally further investigation required for missing product_ids

#### Merge orders_products with products using inner join

Using inner join since 18 products are missing from product table that exists in orders_products_combined

# used inner join to merge prods and orders_products
df_ords_prods_combined_v2 = df_ords_merged.merge(df_prods, on=['product_id'], indicator=True)
df_ords_prods_combined_v2.head()

df_ords_prods_combined_v2.shape

#### Check merge flags

df_ords_prods_combined_v2.value_counts('_merge')

### Export merged dataset

# exporting the dataset as pickle file

df_ords_prods_combined_v2.to_pickle(os.path.join(path, 'Prepared Data/orders_products_merged.pkl'))
