==> 02 Scripts/4. Combining & Export - Part 1.ipynb <==
# Combining & Exporting Datasets

# import libraries

import pandas as pd
import numpy as np
import os

## Orders_Products_Prior

### Import files

# import the orders_checked.csv file and compress the numerics values

path = r'/Users/patel/Documents/CF-Data Anaylst Course/Achievement-4/Instacart Basket Analysis-Oct 2023/02 Data/'
dtypes = {'number_of_orders': 'int8', 'orders_days_of_week': 'int8', 'order_hour_of_day': 'int8','days_since_prior_order': 'float32' }

df_ords = pd.read_csv(os.path.join(path, 'Prepared Data/orders_checked.csv'), dtype=dtypes, index_col = 0)
df_ords.head()

# convert the ids into string object
df_ords['order_id'] = df_ords['order_id'].astype('str')
df_ords['user_id'] = df_ords['user_id'].astype('str')

df_ords.info()

df_ords.shape

# import orders_products_prior.csv and compress numerics values
dtypes = {'add_to_cart_order' : 'int32', 'reordered': 'int8'}
df_ords_prior = pd.read_csv(os.path.join(path, 'Orginal Data/4.3_orders_products/orders_products_prior.csv'), dtype=dtypes, index_col=False)

df_ords_prior.head()

#convert numerics values
df_ords_prior['order_id'] = df_ords_prior['order_id'].astype('str')
df_ords_prior['product_id'] = df_ords_prior['product_id'].astype('str')

df_ords_prior.info()

# descriptive staticts of numerics values
df_ords_prior.describe()

 add_to_cart_order - skewed to the right <br>
 &emsp; min = 1 <br>
 &emsp; max = 145 <br>
 &emsp; mean = 8.35 <br>
 &emsp; 25% = 3 <br>
 &emsp; 50% = 6 <br>
 &emsp; 75% = 11 <br>
 
reordered <br>
 &emsp; min = 25% = 0 <br>
 &emsp; max = 50% = 75% = 1

# dimensions of dataframe
df_ords_prior.shape

### Consistency checks

#### Checking for null

# No null values
df_ords_prior.isnull().sum()

#### Checking for duplicates

df_dups = df_ords_prior[df_ords_prior.duplicated()]
df_dups.shape

df_dups2 = df_ords_prior[df_ords_prior.duplicated(subset=['order_id', 'product_id'], keep=False)]
df_dups2

##### no duplicates or null values found

#### Check value counts

df_ords_prior.value_counts('add_to_cart_order')

df_ords_prior.value_counts('reordered')

1. add_to_cart_order: Total of 145 values. Right-tailed distribution <br>
2. reordered : Total of 2 values (0, 1) -> 59% rows have 1 (more reorders)

## Merge orders and orders_products_prior datasets

#### Inner join used to merge datasets

# using inner join, merge datasets

df_ords_merge = df_ords.merge(df_ords_prior, on=['order_id'], indicator=True)

df_ords_merge.info()

# check number of rows and columns in the merged dataset
df_ords_merge.shape

df_ords_merge.head()

#### Conducting checks (on left merge) 

##### left merged dataset (df_ords_mergeleft) not used. Below code is to investigate null values

# merging using left join on orders with orders_products_prior
df_ords_mergeleft = df_ords.merge(df_ords_prior, on=['order_id'], how='left', indicator=True)

df_ords_mergeleft.info()

# df_ords_mergeleft = 32640698 rows
# df_ords_merge = 32434489

# rows difference = 206209

32640698 - 32434489

# checking null values in mergeleft
df_ords_mergeleft.isnull().sum()

# checking on merge value for each row
df_ords_mergeleft.value_counts('_merge')

# investigating null values.
df_ords_mergeleft[df_ords_mergeleft['product_id'].isnull()]

##### every user has 1 order missing in orders_products_prior

## Exporting merged dataset

df_ords_merge.to_pickle(os.path.join(path, 'Prepared Data/orders_products_combined.pkl'))
