==> 02 Scripts/5. Derive New Variables & Functions.ipynb <==
# Derive New Variables & Functions

# import libraries

import pandas as pd
import numpy as np
import os

## Import Files

# import orders_prods_merged dataset
path = r'/Users/patel/Documents/CF-Data Anaylst Course/Achievement-4/Instacart Basket Analysis-Oct 2023/02 Data/'
df_ords_prod_merged = pd.read_pickle(os.path.join(path, 'Prepared Data/orders_products_merged.pkl'))

df_ords_prod_merged.head()

df_ords_prod_merged.shape

## Subset of Dataset

# extracting first 1,000,000 rows

df = df_ords_prod_merged[:1000000]
df.shape

### Define function and used apply to set price label

# define a function that sets returns price range

def price_label (row):
    if(row['prices'] <= 5):
        return 'Low-Range Product'
    elif (row['prices'] > 5 and row['prices'] <= 15):
        return 'Mid-Range Product'
    elif (row['prices'] > 15):
        return 'High-Range';
    else: 
        return 'Not enough data'

df['price_range'] = df.apply(price_label, axis=1)

df.value_counts('price_range', dropna = False)

### Use loc function to set price label

df.loc[df['prices'] > 15, 'price_range_loc'] = 'High-range product'

df.loc[(df['prices'] <= 15) & (df['prices'] > 5), 'price_range_loc'] = 'Mid-range product'

df.loc[df['prices'] <= 5, 'price_range_loc'] = 'Low-range product'

df['price_range_loc'].value_counts(dropna=False)

### Applying loc function on df_ords_prod_merge

df_ords_prod_merged.loc[df_ords_prod_merged['prices'] > 15, 'price_range_loc'] = 'High-range product'
df_ords_prod_merged.loc[(df_ords_prod_merged['prices'] <= 15) & (df_ords_prod_merged['prices'] > 5), 'price_range_loc'] = 'Mid-range product'
df_ords_prod_merged.loc[df_ords_prod_merged['prices'] <= 5, 'price_range_loc'] = 'Low-range product'

df_ords_prod_merged['price_range_loc'].value_counts(dropna=False)

### Define and set Busiest, Least and Regular days

# orders_days_of_week will be used to define busiest days
df_ords_prod_merged['orders_days_of_week'].value_counts(dropna=False)

results = []

for order in df_ords_prod_merged['orders_days_of_week']:
    if order == 0:
        results.append("Busiest day")
    elif order == 4:
        results.append("Least busy")
    else:
        results.append("Regular busy")

df_ords_prod_merged['busiest_day'] = results

df_ords_prod_merged['busiest_day'].value_counts(dropna=False)

df_ords_prod_merged.head()

## Busiest and Slowest days 

# 0 & 1 busiest days
# 3 & 4 slowest days

results = []

for value in df_ords_prod_merged['orders_days_of_week']:
    if ((value == 0) | (value == 1)):
        results.append("Busiest days")
    elif ((value == 3) | (value == 4)):
        results.append("Slowest days")
    else:
        results.append("Regular days")

df_ords_prod_merged['busy_slow_days'] = results

 ### Checking accuracy of busy_slow_days column

df_ords_prod_merged.value_counts('busy_slow_days')

# counting values for pair orders_days_of_week and busy_slow_days to double check allocation of busy_slow_days values
df_ords_prod_merged.value_counts(['orders_days_of_week', 'busy_slow_days'])


## Busiest hours

### Allocate most, average and fewer orders

#check number of orders per frequecy
df_ords_prod_merged['order_hour_of_day'].value_counts(dropna=False)

#Fewest orders	0,1,2,3,4,5,6
#Average orders	7,8,9,16,17,18,19,20,21,22,23
# Most orders	10,11,12,13,14,15

results = []
fewest_order_hours = [0,1,2,3,4,5]
most_orders_hours = [10,11,12,13,14,15]

for value in df_ords_prod_merged['order_hour_of_day']:
    if (value in fewest_order_hours):
        results.append("Fewest orders")
    elif(value in most_orders_hours):
        results.append("Most orders")
    else:
        results.append("Average orders")

df_ords_prod_merged['busiest_hours'] = results
df_ords_prod_merged.head()

### Check Frequency of busiest hours

df_ords_prod_merged['busiest_hours'].value_counts(dropna=False)

df_ords_prod_merged.value_counts(['order_hour_of_day','busiest_hours'])

### Export to pkl file

df_ords_prod_merged.to_pickle(os.path.join(path, 'Prepared Data/orders_product_merged_busy_dayshours.pkl'))
